fix: rename upgrade plugins to 5.0.0
