--- packages/api-headless-cms/src/plugins/upgrades/v5.0.0-beta.5/deleteCreatedElasticsearchIndices.ts
@@ -1,13 +1,0 @@
-import { CmsContext } from "../../../types";
-import { ElasticsearchConfig } from "../../../utils";
-
-export const deleteCreatedElasticsearchIndices = async (
-    context: CmsContext,
-    indexes: ElasticsearchConfig[]
-): Promise<void> => {
-    await Promise.all(
-        indexes.map(index => {
-            return context.elasticSearch.indices.delete(index);
-        })
-    );
-};
