+++ packages/api-headless-cms/src/plugins/upgrades/v5.0.0/helpers.ts
