+++ packages/app-page-builder/src/admin/plugins/installation.tsx
@@ -122,9 +122,9 @@ export default {
     },
     upgrades: [
         {
-            version: "5.0.0-beta.5",
+            version: "5.0.0",
             getComponent() {
-                return lazy(() => import("./upgrades/v5.0.0-beta.5"));
+                return lazy(() => import("./upgrades/v5.0.0"));
             }
         }
     ]
